Add files via upload
The library is Empty ,so you Will have to fill it. I put 50 arrays for making more memory, first function for adding books data, I made a loop for if 2 books have the same ID it will go out the function and say Error, Second function for Displaying All data and if there no data it will say its empty, and if the book was borrowed it will go out the function and say Error, after that ,fourth function , for returning borrowed books and if is not borrowed it wall go out the f unction and say Error, Finally , Last function Will display the List.